Create allBasin_timeSeries.js
This code is provided as an update to all the one-off codes listed in both the lowerColoradoRegion and lowerColoradoRegion and upperColoradoRegion folders. The code respects the use of the utils.js requirements listed in each of those files. To run specify the index of the watershed on line 63, valid values range from 0-24 (these can be checked by uncommenting line 53). The code is currently set to start with a dry run of the exports, and will print output files without beginning export tasks. To run exports, comment out lines 81-87, and uncomment lines 90-106.
-christian